Undo changes from 0702b22d6598998f97d54b06057b918a9cc24cb6 :
C_Mill_Percent is meant as a switch, to indicate if C_Mill was calculated from a percentage value, or entered as a absolute value.

+++ PipeClassTransfer/Blank.cs
@@ -37,7 +37,7 @@ namespace PipeClassTransfer
         public double C_Mill { get; set; }
         /// <inheritdoc/>
         [System.Xml.Serialization.XmlAttribute]
-        public double C_Mill_Percent { get; set; }
+        public int C_Mill_Percent { get; set; }
         /// <inheritdoc/>
         [System.Xml.Serialization.XmlAttribute]
         public double C_Allowance { get; set; }
